// frontend/src/lib/abis.generated.ts
export const deskAbi = [
	{
		"type": "function",
		"name": "modifyLiquidity",
		"inputs": [
			{
				"name": "key",
				"type": "tuple",
				"internalType": "struct PoolKey",
				"components": [
					{
						"name": "currency0",
						"type": "address",
						"internalType": "Currency"
					},
					{
						"name": "currency1",
						"type": "address",
						"internalType": "Currency"
					},
					{
						"name": "fee",
						"type": "uint24",
						"internalType": "uint24"
					},
					{
						"name": "tickSpacing",
						"type": "int24",
						"internalType": "int24"
					},
					{
						"name": "hooks",
						"type": "address",
						"internalType": "contract IHooks"
					}
				]
			},
			{
				"name": "tickLower",
				"type": "int24",
				"internalType": "int24"
			},
			{
				"name": "tickUpper",
				"type": "int24",
				"internalType": "int24"
			},
			{
				"name": "liquidityDelta",
				"type": "int256",
				"internalType": "int256"
			}
		],
		"outputs": [],
		"stateMutability": "nonpayable"
	},
	{
		"type": "function",
		"name": "msgSender",
		"inputs": [],
		"outputs": [
			{
				"name": "",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "owner",
		"inputs": [],
		"outputs": [
			{
				"name": "",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "pa",
		"inputs": [],
		"outputs": [
			{
				"name": "",
				"type": "address",
				"internalType": "contract IPermissionsAdapter"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "pm",
		"inputs": [],
		"outputs": [
			{
				"name": "",
				"type": "address",
				"internalType": "contract IPoolManager"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "swapToPrice",
		"inputs": [
			{
				"name": "key",
				"type": "tuple",
				"internalType": "struct PoolKey",
				"components": [
					{
						"name": "currency0",
						"type": "address",
						"internalType": "Currency"
					},
					{
						"name": "currency1",
						"type": "address",
						"internalType": "Currency"
					},
					{
						"name": "fee",
						"type": "uint24",
						"internalType": "uint24"
					},
					{
						"name": "tickSpacing",
						"type": "int24",
						"internalType": "int24"
					},
					{
						"name": "hooks",
						"type": "address",
						"internalType": "contract IHooks"
					}
				]
			},
			{
				"name": "targetSqrtPriceX96",
				"type": "uint160",
				"internalType": "uint160"
			},
			{
				"name": "maxIn",
				"type": "uint256",
				"internalType": "uint256"
			}
		],
		"outputs": [],
		"stateMutability": "nonpayable"
	},
	{
		"type": "function",
		"name": "unlockCallback",
		"inputs": [
			{
				"name": "data",
				"type": "bytes",
				"internalType": "bytes"
			}
		],
		"outputs": [
			{
				"name": "",
				"type": "bytes",
				"internalType": "bytes"
			}
		],
		"stateMutability": "nonpayable"
	},
	{
		"type": "function",
		"name": "withdraw",
		"inputs": [
			{
				"name": "token",
				"type": "address",
				"internalType": "contract IERC20"
			},
			{
				"name": "amount",
				"type": "uint256",
				"internalType": "uint256"
			}
		],
		"outputs": [],
		"stateMutability": "nonpayable"
	},
	{
		"type": "error",
		"name": "AddressEmptyCode",
		"inputs": [
			{
				"name": "target",
				"type": "address",
				"internalType": "address"
			}
		]
	},
	{
		"type": "error",
		"name": "AddressInsufficientBalance",
		"inputs": [
			{
				"name": "account",
				"type": "address",
				"internalType": "address"
			}
		]
	},
	{
		"type": "error",
		"name": "FailedInnerCall",
		"inputs": []
	},
	{
		"type": "error",
		"name": "NotOwner",
		"inputs": []
	},
	{
		"type": "error",
		"name": "NotPoolManager",
		"inputs": []
	},
	{
		"type": "error",
		"name": "SafeERC20FailedDecreaseAllowance",
		"inputs": [
			{
				"name": "spender",
				"type": "address",
				"internalType": "address"
			},
			{
				"name": "currentAllowance",
				"type": "uint256",
				"internalType": "uint256"
			},
			{
				"name": "requestedDecrease",
				"type": "uint256",
				"internalType": "uint256"
			}
		]
	},
	{
		"type": "error",
		"name": "SafeERC20FailedOperation",
		"inputs": [
			{
				"name": "token",
				"type": "address",
				"internalType": "address"
			}
		]
	}
] as const;
// ...
export const factoryAbi = [
	{
		"type": "function",
		"name": "createPermissionsAdapter",
		"inputs": [
			{
				"name": "permissionedToken",
				"type": "address",
				"internalType": "contract IERC20"
			},
			{
				"name": "initialOwner",
				"type": "address",
				"internalType": "address"
			},
			{
				"name": "allowListChecker",
				"type": "address",
				"internalType": "contract IAllowlistChecker"
			}
		],
		"outputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "nonpayable"
	},
	{
		"type": "function",
		"name": "permissionsAdapterOf",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		],
		"outputs": [
			{
				"name": "permissionedToken",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "POOL_MANAGER",
		"inputs": [],
		"outputs": [
			{
				"name": "",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "verifiedPermissionsAdapterOf",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		],
		"outputs": [
			{
				"name": "permissionedToken",
				"type": "address",
				"internalType": "address"
			}
		],
		"stateMutability": "view"
	},
	{
		"type": "function",
		"name": "verifyPermissionsAdapter",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		],
		"outputs": [],
		"stateMutability": "nonpayable"
	},
	{
		"type": "event",
		"name": "PermissionsAdapterCreated",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"indexed": true,
				"internalType": "address"
			},
			{
				"name": "permissionedToken",
				"type": "address",
				"indexed": true,
				"internalType": "address"
			}
		],
		"anonymous": false
	},
	{
		"type": "event",
		"name": "PermissionsAdapterVerified",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"indexed": true,
				"internalType": "address"
			},
			{
				"name": "permissionedToken",
				"type": "address",
				"indexed": true,
				"internalType": "address"
			}
		],
		"anonymous": false
	},
	{
		"type": "error",
		"name": "PermissionsAdapterAlreadyVerified",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		]
	},
	{
		"type": "error",
		"name": "PermissionsAdapterNotFound",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		]
	},
	{
		"type": "error",
		"name": "PermissionsAdapterNotVerified",
		"inputs": [
			{
				"name": "permissionsAdapter",
				"type": "address",
				"internalType": "address"
			}
		]
	}
] as const;
// ...
export const eventSignatures: Readonly<Record<string, string>> = {
	"0x031ac7092abd11438885b362e3e4ff3b6999ba04494f91211c7f34cf4d2cb49f": "PermissionsAdapterVerified(address,address)",
	"0x04eddf1a09d47c4653f2d8d8ab3a0e99b462f6032bae11a002b9f919b660c46d": "SwappingEnabledUpdated(bool)",
	"0x06434cb6a9670e5da877306015486047189842c1b3b62f49c8bbf65a1868f54e": "VerificationDeposit(address,uint256)",
	"0x0661566a836c2d45c02e28816c5d4958d7684b29ab0cfcb1025ff7eed3000e7a": "Liquidated(address,address,uint256,uint256,uint256,uint256,uint256)",
	"0x086d215afede6389b8b67581f39e53582a3d2173bae80592af1ebb43af122103": "LiquidationBlockedSet(address,bool)",
	"0x0a3fbbea70e93f2daafa3102f5c9a1b8315e6d7a1e43e4bc020bc1162327470a": "Repaid(address,address,uint256)",
	"0x1b3d7edb2e9c0b0e7c525b20aaaef0f5940d2ed71663c7d39266ecafac728859": "Transfer(address,address,address,uint256,uint256)",
	"0x1e1223c1f6274c6d64fdb08c485e192bd2896c9abf70e840e199ac1f54d6d1d5": "NavUpdated(uint256,uint256)",
	"0x206c478f4d333d6d09c0895132f9aae8dbf95cc1371b76dc3b93072a17265c55": "FlagsSet(address,uint16)",
	"0x29ef05caaff9404b7cb6d1c0e9bbae9eaa7ab2541feba1a9c4248594c08156cb": "Donate(bytes32,address,uint256,uint256)",
	"0x38d16b8cac22d99fc7c124b9cd0de2d3fa1faef420bfe791d8c362d765e22700": "OwnershipTransferStarted(address,address)",
	"0x40db37ff5c0bdc2c427fbb2078c8f24afea940abac0e3c23bb4ea3bf2da2b212": "PausedSet(bool)",
	"0x40e9cecb9f5f1f1c5b9c97dec2917b7ee92e57ba5563708daca94dd84ad7112f": "Swap(bytes32,address,int128,int128,uint160,uint128,int24,uint24)",
	"0x4e83386df19ed27e20e4681a33097114e48ed879ea282347e04ccca610084a6c": "ResidualClaimed(address,uint256)",
	"0x5c2c0d2616a06b35bb159b4d7e227972b59bb33f8d5229ca0e5e438259bfd5d3": "Supplied(address,uint256,uint256)",
	"0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0": "OwnershipTransferred(address,address)",
	"0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925": "Approval(address,address,uint256)",
	"0x92ccf450a286a957af52509bc1c9939d1a6a481783e142e41e2499f0bb66ebc6": "Withdrawn(address,uint256,uint256)",
	"0xac59582e5396aca512fa873a2047e7f4c80f8f55d4a06cb34a78a0187f62719f": "Borrowed(address,uint256)",
	"0xacf055a373fc18efc9064504f586326ebad3223431207428c553428eaf754613": "AllowedWrapperUpdated(address,bool)",
	"0xb3fd5071835887567a0671151121894ddccc2842f1d10bedad13e0d17cace9a7": "Approval(address,address,uint256,uint256)",
	"0xb4bd8ef53df690b9943d3318996006dbb82a25f54719d8c8035b516a2a5b8acc": "ProtocolFeeControllerUpdated(address)",
	"0xbca8d38b73c0bd49fed0184fde2164819be3e78422a9c1ab4a8dab2d10ff6b92": "AllowListCheckerUpdated(address)",
	"0xc30fcfbcaac9e0deffa719714eaa82396ff506a0d0d0eebe170830177288715d": "CollateralWithdrawn(address,uint256)",
	"0xc668126f9ef0f53f8ed212e25aa243fdac9e87eb842f5361ab28b6fd65d67228": "PermissionsAdapterCreated(address,address)",
	"0xceb576d9f15e4e200fdb5096d64d5dfd667e16def20c1eefd14256d8e3faa267": "OperatorSet(address,address,bool)",
	"0xd7243f6f8212d5188fd054141cf6ea89cfc0d91facb8c3afe2f88a1358480142": "CollateralDeposited(address,uint256)",
	"0xdd466e674ea557f56295e2d0218a125ea4b4f0f6f3307b95f85e6110838d6438": "Initialize(bytes32,address,address,uint24,int24,address,uint160,int24)",
	"0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef": "Transfer(address,address,uint256)",
	"0xe9c42593e71f84403b84352cd168d693e2c9fcd1fdbcc3feb21d92b43e6696f9": "ProtocolFeeUpdated(bytes32,uint24)",
	"0xef79f5ec4d2ab65ea4668433504946087afa3662c03b5f51a10a57a0030a60a6": "FrozenSet(address,bool)",
	"0xf208f4912782fd25c7f114ca3723a2d5dd6f3bcc3ac8db5af63baa85f711d5ec": "ModifyLiquidity(bytes32,address,int24,int24,int256,bytes32)",
	"0xf9ec8bbc010f79383d2c837cde44ee91ba7eb3919ae71c38b54964a81963d80f": "ResidualApplied(address,uint256,uint256,uint256)",
	"0xfa8173c933c96f50d089b68a8eaba95a03cbc78d4920a2bd323c11cd01ab50ac": "AllowedHookUpdated(address,bool)",
	"0xfcbc974bf3a532baf2bb229db3c37fd58299b62d2d1db6a855dac5b693bb6ff3": "Liquidated(address,address,uint256,uint256,uint256)"
};